requirements.txt
# Flask Core
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1

# Session Management - Redis Support
Flask-Session==0.5.0
redis==5.0.1

# Caching
Flask-Caching==2.1.0

# Astrology Calculations
pyswisseph==2.10.3.1
numpy==1.24.3
pytz==2023.3

# AI/ML APIs
google-generativeai==0.3.2
openai==1.7.2
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0
aiohttp==3.9.1  # For async HTTP calls (Task 4)
requests-cache==1.2.0

# Database
PyMySQL==1.1.0
cryptography==41.0.7

# Task Scheduling (for cleanup jobs)
APScheduler==3.10.4

# JSON & Data
json5==0.9.14

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Production Server
gunicorn==21.2.0
gevent==23.9.1

# Environment & Config
python-dotenv==1.0.0
